scripts/rcrainfo_salesforce_acount_match/pyproject.toml
[project]
name = "rcrainfo-salesforce-acount-match"
version = "0.1.0"
description = "Salesforce account domain matching script"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "tldextract",
    "python-dotenv",
    "duckdb"
]
